requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
appdirs==1.4.4
args==0.1.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
Babel==2.8.0
bcc==0.18.0
beautifulsoup4==4.13.4
beniget==0.4.1
blinker==1.4
Bottleneck==1.3.2
Brlapi==0.8.3
Brotli==1.1.0
bs4==0.0.2
ccxt==4.4.25
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
chrome-gnome-shell==0.0.0
click==8.0.3
clint==0.5.1
colorama==0.4.4
colorlog==6.9.0
command-not-found==0.3
contourpy==1.3.0
cryptography==43.0.3
cssselect==1.3.0
cssselect2==0.8.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
dlib==19.24.6
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.0.1
exceptiongroup==1.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fake-useragent==2.1.0
filelock==3.20.0
flatbuffers==24.3.25
fonttools==4.54.1
fpdf==1.7.2
frozenlist==1.5.0
fs==2.4.12
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
iniconfig==1.1.1
jdcal==1.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
jsonschema==3.2.0
keras==3.6.0
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
llvmlite==0.38.0
louis==3.20.0
lxml==5.3.1
lxml_html_clean==0.4.2
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
netaddr==0.8.0
netifaces==0.11.0
networkx==3.4.2
numba==0.55.1
numexpr==2.8.1
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
openai-whisper==20240930
opencv-python==4.11.0.86
openpyxl==3.0.9
opt_einsum==3.4.0
optree==0.13.0
optuna==4.0.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parse==1.20.2
pillow==11.0.0
pluggy==0.13.0
ply==3.11
pop-transition==1.1.2
propcache==0.2.0
protobuf==5.28.3
py==1.10.0
pycairo==1.20.1
pycares==4.4.0
pycparser==2.22
pycups==2.0.1
pydbus==0.6.0
pydyf==0.11.0
pyee==11.1.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.2.0
pyphen==0.17.2
pyppeteer==2.0.0
pyquery==2.0.1
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-xlib==0.29
pythran==0.10.0
pytz==2024.2
pyxdg==0.27
PyYAML==6.0.2
regex==2025.10.23
repolib==2.2.1
repoman==1.4.0
requests==2.32.3
requests-html==0.10.0
rich==13.9.3
scikit-learn==1.5.2
scipy==1.14.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.30.0
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spry==0.5.5
SQLAlchemy==2.0.36
sympy==1.14.0
systemd-python==234
ta==0.11.0
tables==3.7.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
texttable==1.6.4
threadpoolctl==3.5.0
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
torch==2.9.0
tqdm==4.66.6
trio==0.29.0
trio-websocket==0.12.2
triton==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
urllib3==2.2.3
uv==0.9.8
w3lib==2.3.1
wadllib==1.3.6
weasyprint==65.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.0
wrapt==1.16.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlwt==1.3.0
yarl==1.17.1
zipp==1.0.0
zopfli==0.2.3.post1
